moved to JPA

+++ trunk/GoJava7/e.singaevskiy/kickstarter_web/src/main/java/ua/com/goit/gojava7/kickstarter/dao/DataSource.java
@@ -6,7 +6,7 @@ public interface DataSource<T> {
     
     List<T> getAll() ;
 
-    T get(int index);
+    T get(Long index);
 
     void add(T element);
